Renamed included cpp files to hpp
