+++ Shared/sdk/UTF8Detect.hpp
